moved initialization from integration to HyperwalletInsight improve threads and javadoc
