+++ commonui/src/main/res/values/strings.xml
@@ -25,5 +25,6 @@
     <string name="transfer_method_list_item_description">Ending on %s</string>
     <string name="not_translated_in_braces">(not translated)</string>
 
-    <string name="environment">PROD</string>
+    <string name="environment_prod">PROD</string>
+    <string name="environment_uat">UAT</string>
 </resources>
